Time: 0 ms (100%), Space: 7.9 MB (86.33%) - LeetHub

+++ 0790-domino-and-tromino-tiling/0790-domino-and-tromino-tiling.cpp
@@ -11,7 +11,7 @@ public:
         int mod = 1e9 + 7;
 
         for (int i = 4; i <= n; i++) {
-            int curr = (c * 2 + a) % mod;
+            int curr = (c * 2 + (long) a) % mod;
             a = b;
             b = c;
             c = curr;
